+++ src/doom/p_floor.cpp
@@ -225,7 +225,7 @@ void T_MoveFloor(floormove_t *floor)
             switch (floor->type)
             {
             case donutRaise:
-                floor->sector->special  = floor->newspecial;
+                floor->sector->special  = static_cast<short>(floor->newspecial);
                 floor->sector->floorpic = floor->texture;
             default:
                 break;
@@ -236,7 +236,7 @@ void T_MoveFloor(floormove_t *floor)
             switch (floor->type)
             {
             case lowerAndChange:
-                floor->sector->special  = floor->newspecial;
+                floor->sector->special  = static_cast<short>(floor->newspecial);
                 floor->sector->floorpic = floor->texture;
             default:
                 break;
@@ -572,13 +572,13 @@ int EV_BuildStairs(line_t *line,
                     continue;
 
                 tsec      = (sec->lines[i])->frontsector;
-                newsecnum = tsec - sectors;
+                newsecnum = static_cast<int>(tsec - sectors);
 
                 if (secnum != newsecnum)
                     continue;
 
                 tsec      = (sec->lines[i])->backsector;
-                newsecnum = tsec - sectors;
+                newsecnum = static_cast<int>(tsec - sectors);
 
                 if (tsec->floorpic != texture)
                     continue;
