+++ src/heretic/r_things.cpp
@@ -86,7 +86,7 @@ void R_InstallSpriteLump(int lump, unsigned frame, unsigned rotation,
         I_Error("R_InstallSpriteLump: Bad frame characters in lump %i", lump);
 
     if (static_cast<int>(frame) > maxframe)
-        maxframe = frame;
+        maxframe = static_cast<int>(frame);
 
     if (rotation == 0)
     {
@@ -102,7 +102,7 @@ void R_InstallSpriteLump(int lump, unsigned frame, unsigned rotation,
         sprtemp[frame].rotate = false;
         for (r = 0; r < 8; r++)
         {
-            sprtemp[frame].lump[r] = lump - firstspritelump;
+            sprtemp[frame].lump[r] = static_cast<short>(lump - firstspritelump);
             sprtemp[frame].flip[r] = static_cast<uint8_t>(flipped);
         }
         return;
@@ -122,7 +122,7 @@ void R_InstallSpriteLump(int lump, unsigned frame, unsigned rotation,
             ("R_InitSprites: Sprite %s : %c : %c has two lumps mapped to it",
              spritename, 'A' + frame, '1' + rotation);
 
-    sprtemp[frame].lump[rotation] = lump - firstspritelump;
+    sprtemp[frame].lump[rotation] = static_cast<short>(lump - firstspritelump);
     sprtemp[frame].flip[rotation] = static_cast<uint8_t>(flipped);
 }
 
@@ -153,12 +153,12 @@ void R_InitSpriteDefs(const char **namelist)
     check = namelist;
     while (*check != nullptr)
         check++;
-    numsprites = check - namelist;
+    numsprites = static_cast<int>(check - namelist);
 
     if (!numsprites)
         return;
 
-    sprites = zmalloc<spritedef_t *>(numsprites * sizeof(*sprites), PU_STATIC, nullptr);
+    sprites = zmalloc<spritedef_t *>(static_cast<unsigned long>(numsprites) * sizeof(*sprites), PU_STATIC, nullptr);
 
     start = firstspritelump - 1;
     end = lastspritelump + 1;
@@ -181,12 +181,12 @@ void R_InitSpriteDefs(const char **namelist)
             {
                 frame = lumpinfo[l]->name[4] - 'A';
                 rotation = lumpinfo[l]->name[5] - '0';
-                R_InstallSpriteLump(l, frame, rotation, false);
+                R_InstallSpriteLump(l, static_cast<unsigned int>(frame), static_cast<unsigned int>(rotation), false);
                 if (lumpinfo[l]->name[6])
                 {
                     frame = lumpinfo[l]->name[6] - 'A';
                     rotation = lumpinfo[l]->name[7] - '0';
-                    R_InstallSpriteLump(l, frame, rotation, true);
+                    R_InstallSpriteLump(l, static_cast<unsigned int>(frame), static_cast<unsigned int>(rotation), true);
                 }
             }
 
@@ -228,9 +228,8 @@ void R_InitSpriteDefs(const char **namelist)
         //
         sprites[i].numframes = maxframe;
         sprites[i].spriteframes =
-            zmalloc<spriteframe_t *>(maxframe * sizeof(spriteframe_t), PU_STATIC, nullptr);
-        memcpy(sprites[i].spriteframes, sprtemp,
-               maxframe * sizeof(spriteframe_t));
+            zmalloc<spriteframe_t *>(static_cast<unsigned long>(maxframe) * sizeof(spriteframe_t), PU_STATIC, nullptr);
+        memcpy(sprites[i].spriteframes, sprtemp, static_cast<unsigned long>(maxframe) * sizeof(spriteframe_t));
     }
 
 }
@@ -315,8 +314,8 @@ vissprite_t *R_NewVisSprite()
         return &overflowsprite;
 
 	numvissprites = numvissprites ? 2 * numvissprites : MAXVISSPRITES;
-	vissprites = static_cast<vissprite_t *>(I_Realloc(vissprites, numvissprites * sizeof(*vissprites)));
-	memset(vissprites + numvissprites_old, 0, (numvissprites - numvissprites_old) * sizeof(*vissprites));
+	vissprites = static_cast<vissprite_t *>(I_Realloc(vissprites, static_cast<unsigned long>(numvissprites) * sizeof(*vissprites)));
+	memset(vissprites + numvissprites_old, 0, (static_cast<unsigned long>(numvissprites - numvissprites_old)) * sizeof(*vissprites));
 
 	vissprite_p = vissprites + numvissprites_old;
 
@@ -848,7 +847,7 @@ void R_SortVisSprites()
     vissprite_t unsorted;
     fixed_t bestscale;
 
-    count = vissprite_p - vissprites;
+    count = static_cast<int>(vissprite_p - vissprites);
 
     unsorted.next = unsorted.prev = &unsorted;
     if (!count)
