+++ src/hexen/r_things.cpp
@@ -105,7 +105,7 @@ void R_InstallSpriteLump(int lump, unsigned frame, unsigned rotation,
         sprtemp[frame].rotate = false;
         for (r = 0; r < 8; r++)
         {
-            sprtemp[frame].lump[r] = lump - firstspritelump;
+            sprtemp[frame].lump[r] = static_cast<short>(lump - firstspritelump);
             sprtemp[frame].flip[r] = static_cast<uint8_t>(flipped);
         }
         return;
@@ -125,7 +125,7 @@ void R_InstallSpriteLump(int lump, unsigned frame, unsigned rotation,
             ("R_InitSprites: Sprite %s : %c : %c has two lumps mapped to it",
              spritename, 'A' + frame, '1' + rotation);
 
-    sprtemp[frame].lump[rotation] = lump - firstspritelump;
+    sprtemp[frame].lump[rotation] = static_cast<short>(lump - firstspritelump);
     sprtemp[frame].flip[rotation] = static_cast<uint8_t>(flipped);
 }
 
@@ -156,7 +156,7 @@ void R_InitSpriteDefs(const char **namelist)
     check = namelist;
     while (*check != nullptr)
         check++;
-    numsprites = check - namelist;
+    numsprites = static_cast<int>(check - namelist);
 
     if (!numsprites)
         return;
@@ -569,7 +569,7 @@ void R_ProjectSprite(mobj_t * thing)
         }
         else
         {
-            vis->clazz = thing->special1.i;
+            vis->clazz = static_cast<int>(thing->special1.i);
         }
         if (vis->clazz > 2)
         {
@@ -851,7 +851,7 @@ void R_SortVisSprites()
     vissprite_t unsorted;
     fixed_t bestscale;
 
-    count = vissprite_p - vissprites;
+    count = static_cast<int>(vissprite_p - vissprites);
 
     unsorted.next = unsorted.prev = &unsorted;
     if (!count)
@@ -990,7 +990,7 @@ void R_DrawSprite(vissprite_t * spr)
     for (x = spr->x1; x <= spr->x2; x++)
     {
         if (clipbot[x] == -2)
-            clipbot[x] = viewheight;
+            clipbot[x] = static_cast<short>(viewheight);
         if (cliptop[x] == -2)
             cliptop[x] = -1;
     }
